+++ lib/features/main_page/features/paged_main_content/presentation/cubit/paged_main_content_cubit.dart
@@ -1,0 +1,8 @@
+import 'package:bloc/bloc.dart';
+import 'package:equatable/equatable.dart';
+
+part 'paged_main_content_state.dart';
+
+class PagedMainContentCubit extends Cubit<PagedMainContentState> {
+  PagedMainContentCubit() : super(PagedMainContentInitial());
+}
