+++ lib/features/main_page/features/paged_main_content/presentation/cubit/paged_main_content_state.dart
@@ -1,0 +1,10 @@
+part of 'paged_main_content_cubit.dart';
+
+abstract class PagedMainContentState extends Equatable {
+  const PagedMainContentState();
+
+  @override
+  List<Object> get props => [];
+}
+
+class PagedMainContentInitial extends PagedMainContentState {}
