+++ lib/features/main_page/features/post/presentation/cubit/post_state.dart
@@ -1,10 +1,46 @@
 part of 'post_cubit.dart';
 
-abstract class PostState extends Equatable {
+@immutable
+abstract class PostState {
   const PostState();
-
-  @override
-  List<Object> get props => [];
 }
 
-class PostInitial extends PostState {}
+class PostInitial extends PostState {
+  const PostInitial();
+}
+
+class PostLoading extends PostState {
+  const PostLoading();
+}
+
+class PostLoaded extends PostState {
+  final Post post;
+
+  const PostLoaded(this.post);
+
+  @override
+  bool operator ==(Object o) {
+    if (identical(this, o)) return true;
+
+    return o is PostLoaded && o.post== post;
+  }
+
+  @override
+  int get hashCode => post.hashCode;
+
+}
+
+class PostError extends PostState {
+  final String message;
+  const PostError(this.message);
+
+  @override
+  bool operator ==(Object o) {
+    if (identical(this, o)) return true;
+
+    return o is PostError && o.message == message;
+  }
+
+  @override
+  int get hashCode => message.hashCode;
+}
