+++ lib/features/main_page/features/post/presentation/widgets/post_widget.dart
@@ -3,105 +3,162 @@
 
 // ignore_for_file: prefer_const_constructors
 
+import 'dart:js';
+
 import 'package:flutter/material.dart';
+import 'package:pahiream_frontend/features/main_page/features/post/data/models/post.dart';
 import 'package:pahiream_frontend/utils/constants.dart';
 import 'package:pahiream_frontend/widgets/global_widgets.dart';
+import 'package:provider/provider.dart';
+
+// var url =
+//     'https://images.unsplash.com/photo-1512820790803-83ca734da794?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxleHBsb3JlLWZlZWR8MTR8fHxlbnwwfHx8fA%3D%3D&auto=format&fit=crop&w=500&q=60';
 
 class PostWidget extends StatelessWidget {
-  const PostWidget({Key? key}) : super(key: key);
+  // TODO implement transforming Post Widget base on data
+  final Post data;
+
+  const PostWidget({
+    Key? key,
+    required this.data,
+  }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-    var url =
-        'https://images.unsplash.com/photo-1512820790803-83ca734da794?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxleHBsb3JlLWZlZWR8MTR8fHxlbnwwfHx8fA%3D%3D&auto=format&fit=crop&w=500&q=60';
     return SizedBox(
       width: 339,
-      child: Card(
-        clipBehavior: Clip.antiAlias,
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(10.0),
-        ),
-        elevation: 16,
-        child: InkWell(
-          onTap: () => print('Pindot card ka HUH'),
-          child: Column(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              PictureAndBanner(url: url),
-              //!Details
-              Padding(
-                padding: const EdgeInsets.fromLTRB(6.0, 10.0, 6.0, 0),
-                child: ListTile(
-                  title: Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                    children: [
-                      PostTitle(),
-                      MouseRegion(
-                        cursor: SystemMouseCursors.click,
-                        child: GestureDetector(
-                          onTap: () => print('Tapped the More Icons'),
-                          child: const Icon(Icons.more_horiz),
-                        ),
-                      )
-                    ],
-                  ),
-                  subtitle: Column(
-                    children: [
-                      SizedBox(height: 10),
-                      Align(alignment: AlignmentDirectional.centerStart, child: Text('PhP 20.00/day')),
-                      PosterRankRow(),
-                      Row(
-                        children: [
-                          Icon(Icons.location_on,
-                              size: 16, color: kFontColorBlack),
-                          SizedBox(width: 5),
-                          Text('7-Eleven'),
-                        ],
-                      ),
-                      Row(
-                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                        children: [
-                          Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: [
-                              Row(
-                                children: [
-                                  Icon(Icons.local_shipping,
-                                      size: 16, color: kFontColorBlack),
-                                  SizedBox(width: 5),
-                                  Text('Cafa, CoS, CiT'),
-                                ],
-                              ),
-                              Row(
-                                children: [
-                                  Icon(Icons.calendar_today,
-                                      size: 16, color: kFontColorBlack),
-                                  SizedBox(width: 5),
-                                  Text('Date posted: '),
-                                ],
-                              ),
-                            ],
+      child: Provider(
+        //  TODO: Not yet tested if proper syntax
+        create: (context) => data,
+        child: Card(
+          clipBehavior: Clip.antiAlias,
+          shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(10.0),
+          ),
+          elevation: 16,
+          child: InkWell(
+            onTap: () => print('Pindot card ka HUH'),
+            child: Column(
+              mainAxisSize: MainAxisSize.min,
+              children: [
+                PictureAndBanner(url: data.image_location),
+                //!Details
+                Padding(
+                  padding: const EdgeInsets.fromLTRB(6.0, 10.0, 6.0, 0),
+                  child: ListTile(
+                    title: Row(
+                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                      children: [
+                        PostTitle(),
+                        MouseRegion(
+                          cursor: SystemMouseCursors.click,
+                          child: GestureDetector(
+                            onTap: () => print('Tapped the More Icons'),
+                            child: const Icon(Icons.more_horiz),
                           ),
-                          MouseRegion(
-                            cursor: SystemMouseCursors.click,
-                            child: GestureDetector(
-                                onTap: () {
-                                  print('\n\nSino kakausapin mo gagu');
-                                },
-                                child: Icon(Icons.mode_comment_outlined)),
-                          )
-                        ],
-                      )
-                    ],
+                        )
+                      ],
+                    ),
+                    subtitle: Column(
+                      children: [
+                        SizedBox(height: 10),
+                        SecondRowBuilder(),
+                        PosterRankRow(),
+                        Row(
+                          children: [
+                            Icon(Icons.location_on,
+                                size: 16, color: kFontColorBlack),
+                            SizedBox(width: 5),
+                            Text('7-Eleven'),
+                          ],
+                        ),
+                        Row(
+                          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                          children: [
+                            LocationDestinnationRowBuilder(),
+                            MouseRegion(
+                              cursor: SystemMouseCursors.click,
+                              child: GestureDetector(
+                                  onTap: () {
+                                    print('\n\nSino kakausapin mo gagu');
+                                  },
+                                  child: Icon(Icons.mode_comment_outlined)),
+                            )
+                          ],
+                        )
+                      ],
+                    ),
                   ),
                 ),
-              ),
-              //ButtonBAr
-            ],
+                //ButtonBAr
+              ],
+            ),
           ),
         ),
       ),
     );
+  }
+}
+
+class LocationDestinnationRowBuilder extends StatelessWidget {
+  const LocationDestinnationRowBuilder({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    Post data = Provider.of<Post>(context);
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        LocationDestinationBuilder(),
+        Row(
+          children: [
+            Icon(Icons.calendar_today, size: 16, color: kFontColorBlack),
+            SizedBox(width: 5),
+            Text(data.date),
+          ],
+        ),
+      ],
+    );
+  }
+}
+
+// TODO: gawin Widget nito 01/06/2022
+class LocationDestinationBuilder extends StatelessWidget {
+  const LocationDestinationBuilder({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    Post data = Provider.of<Post>(context);
+    bool isPahiram = data.rent_due.isNotEmpty; //note: pahiram lang ang may rent_due
+    return Row(
+      children: [
+        Icon(Icons.local_shipping, size: 16, color: kFontColorBlack),
+        SizedBox(width: 5),
+// TODO: gawin Widget nito 01/06/2022
+        Text(data.delivery_time),
+      ],
+    );
+  }
+}
+
+// TODO: gawin Widget nito 01/06/2022
+class SecondRowBuilder extends StatelessWidget {
+  const SecondRowBuilder({
+    Key? key,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    Post data = Provider.of<Post>(context);
+    bool isPahiram =
+        data.rent_due.isNotEmpty; //note: pahiram lang ang may rent_due
+    return Align(
+        alignment: AlignmentDirectional.centerStart,
+        child: Text('PhP 20.00/day'));
   }
 }
 
@@ -112,15 +169,25 @@ class PosterRankRow extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    List<String> ranks = ['Basic', 'Intermediate', 'Savior'];
+    Post data = Provider.of<Post>(context);
+    bool isPahiram =
+        data.rent_due.isNotEmpty; //note: pahiram lang ang may rent_due
+    String rank = int.parse(data.points) < 50
+        ? ranks.elementAt(0)
+        : (int.parse(data.points) > 50 && int.parse(data.points) < 100)
+            ? ranks.elementAt(1)
+            : ranks.elementAt(2);
+
     return Row(
       children: [
         Icon(Icons.person, size: 16, color: kFontColorBlack),
         SizedBox(width: 5),
-        Text('Uzumaki Naruto'),
+        Text(data.first_name + " " + data.last_name),
         SizedBox(width: 5),
         Icon(Icons.fiber_manual_record, size: 16, color: Colors.blue),
         SizedBox(width: 5),
-        Text('Savior', style: CommonStyleText.txtStyle(color: kPrimaryPink)),
+        Text(rank, style: CommonStyleText.txtStyle(color: kPrimaryPink)),
       ],
     );
   }
@@ -133,8 +200,12 @@ class PostTitle extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    Post data = Provider.of<Post>(context);
+    bool isPahiram =
+        data.rent_due.isNotEmpty; //note: pahiram lang ang may rent_due
+    String title = isPahiram ? data.item : data.title;
     return Text(
-      'Physics Book Vol. 1',
+      title,
       style: CommonStyleText.txtStyle(size: 16, weigth: bold),
     );
   }
@@ -150,6 +221,7 @@ class PictureAndBanner extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    Post data = Provider.of<Post>(context);
     return Stack(
       // !Picture
       alignment: Alignment.topRight,
@@ -160,7 +232,8 @@ class PictureAndBanner extends StatelessWidget {
             alignment: Alignment.center,
             decoration: BoxDecoration(
                 image: DecorationImage(
-                    image: NetworkImage(url), fit: BoxFit.fitWidth))),
+                    image: NetworkImage(data.image_location),
+                    fit: BoxFit.fitWidth))),
         Banner()
       ],
     );
@@ -174,14 +247,24 @@ class Banner extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    Post data = Provider.of<Post>(context);
+    bool isPahiram =
+        data.rent_due.isNotEmpty; //note: pahiram lang ang may rent_due
+    Color bannerColor = (((!isPahiram) && data.type == 'delivery')
+            ? kBanner1
+            : ((!isPahiram) && data.type == 'request')
+                ? kBanner2
+                : ((isPahiram) && data.type == 'rent')
+                    ? kBanner3
+                    : kBanner4
+        // ((isPahiram) && data.type=='request' ) ? Kbanner4
+        );
     return Container(
       decoration: null,
-      // ! Banner Color
-      color: const Color.fromARGB(255, 221, 114, 107),
+      color: bannerColor,
       padding: const EdgeInsets.all(5.0),
-      // ! Banner Text
       child: Text(
-        'for rent'.toUpperCase(),
+        data.type.toUpperCase(),
         style: CommonStyleText.txtStyle(
             weigth: bold, color: kFontColorWhite, size: 14),
       ),
