+++ lib/utils/constants.dart
@@ -23,6 +23,12 @@ Color kFontColorRed = const Color(0xffC41E3A);
 Color kFontColorRedWarning = const Color(0xffE70E0E);
 Color kGray = const Color(0xffDEDEE1);
 
+// Banner Colors
+Color kBanner1 = Color.fromARGB(255, 58, 6, 202);
+Color kBanner2 = Color.fromARGB(255, 231, 119, 14);
+Color kBanner3 = Color.fromARGB(255, 230, 64, 23);
+Color kBanner4 = Color.fromARGB(255, 47, 231, 14);
+
 // Font Size
 double kFontSizeHeader = pToF(48);
 double kFontSizeSubtitle = pToF(18);
