+++ pubspec.yaml
@@ -40,6 +40,7 @@ dependencies:
   bloc: ^8.0.2
   equatable: ^2.0.3
   flutter_bloc: ^8.0.1
+  provider: ^6.0.2
 
 dev_dependencies:
   flutter_test:
